+++ packages/reactive_ble_mobile/android/src/main/kotlin/com/signify/hue/flutterreactiveble/ble/ReactiveBleClient.kt
@@ -1,5 +1,8 @@
 package com.signify.hue.flutterreactiveble.ble
 
+import android.Manifest
+import android.annotation.SuppressLint
+import android.annotation.TargetApi
 import android.bluetooth.BluetoothDevice.BOND_BONDING
 import android.bluetooth.BluetoothGattCharacteristic
 import android.content.Context
@@ -38,7 +41,11 @@ import android.bluetooth.BluetoothAdapter
 import android.bluetooth.BluetoothManager
 import android.bluetooth.le.*
 import android.bluetooth.*
+import android.bluetooth.BluetoothDevice
+import android.content.pm.PackageManager
 import android.util.Log
+import androidx.core.app.ActivityCompat
+import com.signify.hue.flutterreactiveble.utils.BuildConfig
 import kotlin.random.Random
 
 private const val tag: String = "ReactiveBleClient"
@@ -248,7 +255,7 @@ open class ReactiveBleClient(private val context: Context) : BleClient {
                             the bonding with the peripheral. By retrying the operation once we
                             deviate between this flaky one time error and real auth failed cases
                             */
-                            .retry(1) { Build.VERSION.SDK_INT < Build.VERSION_CODES.O }
+                            .retry(1) { BuildConfig().getVersionSDKInt() < Build.VERSION_CODES.O }
                             .map { value ->
                                 CharOperationSuccessful(deviceId, value.asList())
                             }
@@ -755,7 +762,7 @@ open class ReactiveBleClient(private val context: Context) : BleClient {
                         null
                     );
 
-                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+                    if (BuildConfig().getVersionSDKInt() >= Build.VERSION_CODES.M) {
                         mBluetoothGatt = device?.connectGatt(
                             context,
                             false,
@@ -979,9 +986,9 @@ open class ReactiveBleClient(private val context: Context) : BleClient {
             return
         }
         while (!mBluetoothGattServer!!.removeService(service)) {
-            Log.d(tag, "gatt service ${service} not removed")
+            Log.d(tag, "gatt service $service not removed")
         }
-        Log.d(tag, "gatt service ${service} removed")
+        Log.d(tag, "gatt service $service removed")
     }
 
     override fun addGattService() {
@@ -995,20 +1002,72 @@ open class ReactiveBleClient(private val context: Context) : BleClient {
     }
 
     /**
+     * Checks if a bonded device is of type classic bonding. This can be determined by the device type.
+     * The device types value can be [BluetoothDevice.DEVICE_TYPE_CLASSIC] or [BluetoothDevice.DEVICE_TYPE_DUAL]
+    */
+    @TargetApi(34)
+    fun isClassicBonding(device: BluetoothDevice, ctx: Context?, buildCnfg: BuildConfig?): Boolean {
+        try {
+            val config = buildCnfg ?: BuildConfig()
+            val btPermission = when (config.getVersionSDKInt()) {
+                in 1..Build.VERSION_CODES.R -> Manifest.permission.BLUETOOTH
+                else -> Manifest.permission.BLUETOOTH_CONNECT
+            }
+            if (ActivityCompat.checkSelfPermission(
+                    ctx ?: context,
+                    btPermission
+                ) != PackageManager.PERMISSION_GRANTED
+            ) {
+                Log.i(tag, "Bt permission denied");
+                return false
+            }
+            return (device.getType() == BluetoothDevice.DEVICE_TYPE_CLASSIC) || (device.getType() == BluetoothDevice.DEVICE_TYPE_DUAL)
+        } catch (e: Exception) {
+            Log.i(tag, "An error occured while checking the iNet System bonding type")
+            Log.i(tag, e.toString())
+            return false;
+        }
+    }
+
+    /**
      * Checks if old bonding exists, and if, allows showing the delete iNetBox Bondings pop up.
      * */
+    @TargetApi(34)
     override fun checkIfOldInetBoxBondingExists(deviceId: String): Boolean {
         try {
             val bluetoothManager: BluetoothManager =
                 ctx.getSystemService(Context.BLUETOOTH_SERVICE) as BluetoothManager
             val bluetoothAdapter: BluetoothAdapter = bluetoothManager.adapter
 
-            Log.i(tag, "check bonding for device id: ${deviceId}");
+            Log.i(tag, "check bonding for device id: $deviceId");
+
+
+            val btPermission =
+                when (com.signify.hue.flutterreactiveble.utils.BuildConfig().getVersionSDKInt()) {
+                    in 1..Build.VERSION_CODES.R -> Manifest.permission.BLUETOOTH
+                    else -> Manifest.permission.BLUETOOTH_CONNECT
+                }
+            /*if (BuildConfig().checkSelfPermission(
+                    context,
+                    btPermission
+                ) != PackageManager.PERMISSION_GRANTED
+            ) {
+                Log.i(tag, "Bluetooth permission not granted!");
+                return false
+            }*/
+            if (ActivityCompat.checkSelfPermission(
+                    context,
+                    btPermission
+                ) != PackageManager.PERMISSION_GRANTED
+            ) {
+                Log.i(tag, "Bluetooth permission not granted!");
+                return false
+            }
 
             val bondedDevices: Set<BluetoothDevice> = bluetoothAdapter.getBondedDevices()
-
-            if (searchForBondedDevice(deviceId, bondedDevices) != null) {
-                Log.i(tag, "found iNet Box bonding")
+            val bondedDevice = searchForBondedDevice(deviceId, bondedDevices)
+            if (bondedDevice != null && isClassicBonding(bondedDevice, null, null)) {
+                Log.i(tag, "found old iNet Box bonding")
                 return true;
             }
             Log.i(tag, "found no iNet Box bonding")
@@ -1026,32 +1085,65 @@ open class ReactiveBleClient(private val context: Context) : BleClient {
      * If the bonded iNetBox is from type 0x00000002 - DEVICE_TYPE_LE this iNetBox was
      * bonded via the new flutter app.
      * https://developer.android.com/reference/android/bluetooth/BluetoothDevice#DEVICE_TYPE_DUAL
+     * If [forceDelete] is set, the bonded device will be removed regardless of the device type.
      * */
-    override fun removeInetBoxBonding(deviceId: String): Boolean {
+    override fun removeInetBoxBonding(
+        deviceId: String,
+        forceDelete: Boolean,
+        context: Context?,
+    ): Boolean {
         try {
-            val bluetoothManager: BluetoothManager =
+            val ctx = context ?: this.context
+            val btManager: BluetoothManager =
                 ctx.getSystemService(Context.BLUETOOTH_SERVICE) as BluetoothManager
-            val bluetoothAdapter: BluetoothAdapter = bluetoothManager.adapter
 
+            val bluetoothAdapter: BluetoothAdapter = btManager.adapter
+
+            val btPermission = when (BuildConfig().getVersionSDKInt()) {
+                in 1..Build.VERSION_CODES.R -> Manifest.permission.BLUETOOTH
+                else -> Manifest.permission.BLUETOOTH_CONNECT
+            }
+            if (ActivityCompat.checkSelfPermission(
+                    ctx,
+                    btPermission
+                ) != PackageManager.PERMISSION_GRANTED
+            ) {
+                Log.i(tag, "Bt permission denied");
+                return false
+            }
+
+            Log.i(tag, "Bluetooth permission not granted!")
             val bondedDevices: Set<BluetoothDevice> = bluetoothAdapter.getBondedDevices()
 
-            Log.i(tag, "start remove bonding for device id: ${deviceId}");
+            Log.i(tag, "start remove bonding for device id: $deviceId");
 
             val foundDevice: BluetoothDevice? = searchForBondedDevice(deviceId, bondedDevices)
-            if (foundDevice != null) {
+
+            if (foundDevice == null) {
+                Log.i(tag, "Found no related device to remove bonding")
+                return false;
+            }
+
+            val isBtClassicBonding: Boolean = isClassicBonding(foundDevice, null, null)
+
+            Log.i(tag, "is classic bonding: $isBtClassicBonding")
+            Log.i(tag, "forceDelete: $forceDelete")
+
+            if (isBtClassicBonding || forceDelete) {
                 Log.i(tag, "found iNet Box bonding - remove bond")
                 val pair = foundDevice.javaClass.getMethod("removeBond")
                 pair.invoke(foundDevice)
                 Log.i(tag, "remove iNet Box bonding - success")
                 return true;
             }
-            Log.i(tag, "remove iNet Box bonding - failure")
+            Log.i(tag, "No classic bonding found, neither forcing deletion.")
             return false;
         } catch (e: Exception) {
-            Log.i(tag, "Error removing bonding: " + e);
+            Log.i(tag, "Error removing bonding: $e");
             return false;
         }
     }
+
 
     /**
      * [deviceId] is empty, in case of an existing bonding with the iNet Box, paired through the
@@ -1065,30 +1157,33 @@ open class ReactiveBleClient(private val context: Context) : BleClient {
      * Android onto bluetooth low energy. This action helps to clean up the cached bluetooth
      * services after a firmware migration.
      * */
-    private fun searchForBondedDevice(
+    @SuppressLint("MissingPermission")
+    fun searchForBondedDevice(
         deviceId: String,
         bondedDevices: Set<BluetoothDevice>
     ): BluetoothDevice? {
         try {
             for (device in bondedDevices) {
-                var deviceName: String? = device.getName()
+
+                val deviceName: String? = device.name
                 Log.i(tag, "found device");
                 Log.i(tag, deviceName.toString());
-                Log.i(tag, device.getType().toString());
+                Log.i(tag, device.type.toString());
                 Log.i(tag, device.toString());
-                Log.i(tag, device.getAddress().toString());
+                Log.i(tag, device.address.toString());
                 val isSearchedDevice: Boolean =
                     (device.getAddress() == deviceId) || (deviceName ?: "").contains("iNet Box")
-                if ( isSearchedDevice && ((device.getType() == 0x00000001) || (device.getType() == 0x00000003))) {
+
+                if (isSearchedDevice) {
                     Log.i(tag, "found bonded iNet Box");
                     return device;
                 }
-                Log.i(tag, "no bonded iNet Box found");
             }
+            Log.i(tag, "No existing bonding found!");
             return null;
         } catch (e: Exception) {
             Log.e(tag, "Exception while iterating over bondings");
-            Log.e(tag, "Exception: ${e.toString()}");
+            Log.e(tag, "Exception: $e");
             return null;
         }
     }
@@ -1101,8 +1196,8 @@ open class ReactiveBleClient(private val context: Context) : BleClient {
         // TODO write to local characteristic and notify
         val servicesList: List<BluetoothGattService> = mBluetoothGattServer!!.getServices()
 
-        for (i in 0 until servicesList.size) {
-            val bluetoothGattService: BluetoothGattService = servicesList[i]
+        for (element in servicesList) {
+            val bluetoothGattService: BluetoothGattService = element
 
             val bluetoothGattCharacteristicList: List<BluetoothGattCharacteristic> =
                 bluetoothGattService.getCharacteristics()
@@ -1163,7 +1258,7 @@ open class ReactiveBleClient(private val context: Context) : BleClient {
                         the bonding with the peripheral. By retrying the operation once we
                         deviate between this flaky one time error and real auth failed cases
                         */
-                        .retry(1) { Build.VERSION.SDK_INT < Build.VERSION_CODES.O }
+                        .retry(1) { BuildConfig().getVersionSDKInt() < Build.VERSION_CODES.O }
                         .map { value ->
                             CharOperationSuccessful(deviceId, value.asList())
                         }